backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
python-dotenv>=1.0
httpx>=0.28
aiosqlite>=0.21
aiosmtplib>=3.0
apscheduler>=3.10
pydantic>=2.10
dnspython>=2.7
